Fixed infinite loop in Rust code generation for some recursive types
In writing cross-language tests, I found that Rust code generation would loop forever when a recursive type was referred to from another type. Turns out that this was a bug in the `refersTo` function—rewriting `refersTo` to keep track of the types it's already seen and avoid cycles fixed the problem.