Fixed how Name and ModuleName are displayed in error messages
Ideally, I'd like to change the `p` quasiquoter to use the `Pretty` class directly, but that looks like it will be difficult—I don't see any way of doing that on top of `string-interpolate` without modifying the library code directly.

Instead, I just changed the `Show` instances for both `Name` and `ModuleName` to use `Pretty`. To make this code a bit cleaner, I added a newtype `ShowPretty` that can be used with `DerivingVia`:

``` haskell
data MyType = MyType {...}
  deriving Show via ShowPretty MyType
```

Deriving via is pretty cool!