Removed Paths_theta module from lib: and exe: components
Depending on the autogenerated `Paths_theta` module causes unnecessary rebuilds on Stack (see #37). Depending on `Paths_theta` from the test suite does not cause the same problem since the project can be configured with tests disabled.

Since we only used `Paths_theta` in the library to get the package version (in `Theta.Version`), the easiest solution was to hardcode the version in the library and use a unit test to make sure it's up to date. This is a bit hacky and manual, but the unit test will keep me from forgetting to update the hardcoded version so it isn't a big deal in practice.