Added the UUID primitive type
Added the `UUID` type to Theta (given `language-version` ≥ 1.1.0). The UUID type:

  * compiles to `{ "type": "string", "logicalType" : "uuid" }` in Avro
  * [`UUID`][haskell-type] from [`uuid`][haskell-package] in Haskell
  * [`UUID`][python-type] in Python
  * [`Uuid`][rust-type] from [`uuid`][rust-package] in Rust
  * [`UUID`][kotlin-type] in Kotlin

The refactoring in #51 definitely helped—I could follow type errors and incomplete-pattern-matching warnings to find most of the places I needed to update in the code. Most of the remaining work ended up in updating tests and fixing the errors they exposed.

[haskell-type]: https://hackage.haskell.org/package/uuid-1.3.15/docs/Data-UUID.html#t:UUID
[haskell-package]: https://hackage.haskell.org/package/uuid
[python-type]: https://docs.python.org/3/library/uuid.html#uuid.UUID
[rust-type]: https://docs.rs/uuid/latest/uuid/struct.Uuid.html
[rust-package]: https://docs.rs/uuid/latest/uuid/
[kotlin-type]: https://docs.oracle.com/javase/7/docs/api/java/util/UUID.html