Add Time primitive type
This PR adds a `Time` primitive type (given `language-version` ≥ 1.1.0). The `Time` type represents a time of day (ie time since midnight) with no date or timezone. It compiles to:

  * `{ "type" : "long", "logicalType" : "time-micros" }` in Avro
  * `DayTime` in Haskell: a newtype over [`DiffTime`][DiffTime] defined in `Theta.Value`[^haskell]
  * [`datetime.time`][python-time] in Python
  * [`NaiveTime`][rust-time] from [chrono] in Rust
  * [`LocalTime`][kotlin-time] in Kotlin

[^haskell]: Haskell's [`time`][haskell-time] library does not have a dedicated "time of day" type, but I want to reserve `DiffTime` for adding a `Duration` primitive type, so I wrapped `DiffTime` in a newtype.

I will likely implement a `LocalDateTime` soon, corresponding to Avro's [`local-timestamp-micros`][avro-local-datetime] logical type. I also want to implement a `Duration` type, but since that's [encoded in Avro][avro-duration] as a fixed-size type, it'll need some support work first.

I currently do not plan to implement Avro's millisecond-precision variations on `time`, `timestamp` and `local-timestamp`, but that might change if I have a good idea about how to do that. (Maybe with annotations or something?)

[DiffTime]: https://hackage.haskell.org/package/time/docs/Data-Time-Clock.html#t:DiffTime

[haskell-time]: https://hackage.haskell.org/package/time

[python-time]: https://docs.python.org/3/library/datetime.html#datetime.time

[rust-time]: https://docs.rs/chrono/0.4.0/chrono/naive/struct.NaiveTime.html

[kotlin-time]: https://docs.oracle.com/javase/8/docs/api/java/time/LocalTime.html

[avro-local-datetime]: https://avro.apache.org/docs/current/spec.html#Local+timestamp+%28microsecond+precision%29

[avro-duration]: https://avro.apache.org/docs/current/spec.html#Duration