Fix Time implementation
After merging #56, I found that [time] *did* have a local-time type called [`TimeOfDay`][time-of-day], so this PR replaces my `DayTime` newtype with that instead.

I also forgot to add the `Time` type to `test/data/modules/primitives.theta`, which hid a couple of other bugs. Now the test module is updated and the bugs are fixed.