Feat: Reader schemas default to null for optionals
Some related discussions in: <https://github.com/target/theta-idl/issues/70>

Theta is not generating a `default` specification in record fields that
are specified as optional.

This meant that readers (using the haskell `avro` library) could not
decode data, failing with a message like

```
AvroDecodingException "No default found for deconflicted field [someOptionalField]"
```

The specification says:

> Default values for union fields correspond to the first schema in the union.

From: <https://avro.apache.org/docs/current/spec.html#Unions>

It is unclear whether this part of the specification is meant to be
implemented or not.

Now `theta` will take an approach where Optional types will carry
`default` specification that defaults the value to null in the reader
when it's missing in the writer schema.